+++ Kokoro/docs/release-notes.md
@@ -5,6 +5,7 @@
 ## Upcoming release
 
 * Fixed Limited/Finite traits and Sequence actions sometimes getting garbled icons (related to screen shake).
+* Improved tooltip wording of Impulsive, On Discard and On Turn End.
 
 ## 2.9.0
 Released 25 April 2025.